Move the tests into a submodule
